[sw,hello_world] Split off gpio_demo from hello_world
Make `hello_world` much simpler by having it just initialize the UART
and just print "Hello World!". Rename the original `hello_world` to
`gpio_demo`.
