+++ sw/device/examples/hello_world/BUILD
@@ -2,57 +2,28 @@
 # Licensed under the Apache License, Version 2.0, see LICENSE for details.
 # SPDX-License-Identifier: Apache-2.0
 
-load("//rules/opentitan:defs.bzl", "opentitan_binary")
-load("//rules:files.bzl", "exclude_files")
-load("@rules_pkg//pkg:mappings.bzl", "pkg_files")
+load("//rules/opentitan:defs.bzl", "opentitan_test")
 
 package(default_visibility = ["//visibility:public"])
 
-opentitan_binary(
+opentitan_test(
     name = "hello_world",
-    testonly = True,
     srcs = ["hello_world.c"],
     copts = [
         # Disable the date-time warning only for the hello world program.
         "-Wno-date-time",
     ],
-    exec_env = [
-        "//hw/top_earlgrey:fpga_cw310",
-        "//hw/top_earlgrey:fpga_cw340",
-        "//hw/top_earlgrey:sim_dv",
-        "//hw/top_earlgrey:sim_verilator",
-        # For some reason, this target makes englishbreakfast builds fail.
-        #"//hw/top_earlgrey:silicon_owner_sival_rom_ext",
-    ],
+    exec_env = {
+        "//hw/top_earlgrey:sim_dv": None,
+        "//hw/top_earlgrey:sim_verilator": None,
+    },
     deps = [
-        ":hello_world_lib",
-        "//sw/device/lib/base:mmio",
-    ],
-)
-
-cc_library(
-    name = "hello_world_lib",
-    deps = [
-        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/examples:demos",
+	"//hw/top/dt",
         "//sw/device/lib/arch:device",
-        "//sw/device/lib/crt",
-        "//sw/device/lib/dif:gpio",
-        "//sw/device/lib/dif:pinmux",
         "//sw/device/lib/dif:uart",
-        "//sw/device/lib/runtime:hart",
-        "//sw/device/lib/runtime:log",
         "//sw/device/lib/runtime:print_uart",
-        "//sw/device/lib/testing:pinmux_testutils",
         "//sw/device/lib/testing/test_framework:check",
         "//sw/device/lib/testing/test_framework:ottf_start",
         "//sw/device/lib/testing/test_framework:ottf_test_config",
     ],
-)
-
-pkg_files(
-    name = "package",
-    testonly = True,
-    srcs = [":hello_world"],
-    prefix = "earlgrey/examples",
 )
